+++ CHANGELOG.rst
@@ -64,6 +64,8 @@ in development
   functionality (file upload, directory upload, etc.) is not used. (improvement)
   Reported by  Cody A. Ray
 * API and CLI allow rules to be filtered by their enable state. (improvement)
+* Fix SSH bastion host support by ensuring the bastion parameter is passed to the paramiko ssh
+  client. (bug-fix) #2543 [Adam Mielke]
 
 1.4.0 - April 18, 2016
 ----------------------